sentiment-aura/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
openai>=2.7.0
pydantic>=2.9.0
websockets>=12.0
httpx>=0.21.0

